all components integrated
All components are now integrated with the local initiative files and
graph instance.

The major UX shortcomings right now are the interpolation of user nodes
from url keys, and the formatting of the text displayed in the node
dropdowns. Once we get the Reference detector up and running, the
formatting should be easier to sort out. However, it looks like reverse
resolutions might not be supported yet in sourceCred core, where a link
gets passed in and an alias that composes a username can be resolved
for. It appears that only the username itself can be easily resolved
right now